Move pathfinding algorithm to its own folder
