+++ lib/dijkstra/package.yml
@@ -1,0 +1,8 @@
+include_directories:
+  - include
+
+depends:
+  - test-runner
+
+tests:
+  - tests/dijkstra.cpp
