+++ master-firmware/package.yml
@@ -2,6 +2,7 @@ dependency-dir: ../lib
 
 depends:
     - crc
+    - dijkstra
     - parameter
     - nanopb
     - pid
@@ -116,7 +117,6 @@ tests:
     - tests/msgbus_protobuf.cpp
     - tests/test_3dof_pendulum_kinematics.cpp
     - tests/test_manipulator_control.cpp
-    - tests/test_arm_pathfinding.cpp
 
 templates:
     app_src.mk.jinja: app_src.mk
